requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.13.1
pydantic==2.5.0
redis==5.0.1
boto3==1.34.0
kubernetes==28.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pydantic-settings==2.1.0
minio==7.2.0
requests==2.31.0
groq==0.4.1
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
alembic>=1.10.0